Port base_usability, sale_usability, sale_stock_usability to v10
