+++ sale_stock_usability/sale_stock_view.xml
@@ -1,26 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <!--
-    Copyright (C) 2015 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
+  © 2015-2016 Akretion (http://www.akretion.com/)
+  @author Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data>
+<odoo>
 
-
-<record id="view_picking_internal_search_inherit" model="ir.ui.view">
-    <field name="name">sale_stock_usability.picking.search</field>
-    <field name="model">stock.picking</field>
-    <field name="inherit_id" ref="sale_stock.view_picking_internal_search_inherit" />
-    <field name="arch" type="xml">
-        <filter name="to_invoice" position="attributes">
-            <attribute name="domain">[('invoice_state', '=', '2binvoiced'), ('state', '=', 'done')]</attribute>
-        </filter>
-    </field>
-</record>
-
+<!-- the truck icon is probably enough
 <record id="view_order_form_inherit" model="ir.ui.view">
     <field name="name">sale_stock_usability.sale_order_form</field>
     <field name="model">sale.order</field>
@@ -33,23 +21,6 @@
         </notebook>
     </field>
 </record>
+-->
 
-
-<record id="procurement_group_form_view" model="ir.ui.view">
-    <field name="name">sale_stock_usability.procurement.group.form</field>
-    <field name="model">procurement.group</field>
-    <field name="inherit_id" ref="procurement.procurement_group_form_view"/>
-    <field name="arch" type="xml">
-        <xpath expr="//field[@name='move_type']/.." position="after">
-            <group name="sale" string="Sale Orders">
-                <field name="sale_ids" nolabel="1"/>
-            </group>
-            <group name="picking" string="Pickings">
-                <field name="picking_ids" nolabel="1"/>
-            </group>
-        </xpath>
-    </field>
-</record>
-
-</data>
-</openerp>
+</odoo>
