+++ sale_usability/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
 
 from . import sale
-from . import account_invoice
