+++ sale_usability/sale_view.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-   Copyright (C) 2015 Akretion (http://www.akretion.com/)
-   @author: Alexis de Lattre <alexis.delattre@akretion.com>
-   The licence is in the file __openerp__.py
+  © 2015-2016 Akretion (http://www.akretion.com/)
+  @author: Alexis de Lattre <alexis.delattre@akretion.com>
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 
-<openerp>
-<data>
+<odoo>
 
 <record id="view_order_form" model="ir.ui.view">
     <field name="name">usability.sale.order.form</field>
@@ -18,11 +17,7 @@
                 <field name="invoice_ids" nolabel="1" context="{'form_view_ref': 'account.invoice_form'}"/>
             </page>
         </notebook>
-        <field name="date_order" position="after">
-            <field name="date_confirm"
-                attrs="{'invisible': [('state', 'in', ('draft', 'sent'))]}"/>
-        </field>
-        <field name="fiscal_position" position="attributes">
+        <field name="fiscal_position_id" position="attributes">
             <attribute name="widget">selection</attribute>
         </field>
     </field>
@@ -34,10 +29,7 @@
     <field name="inherit_id" ref="sale.view_quotation_tree"/>
     <field name="arch" type="xml">
         <field name="amount_total" position="before">
-            <field name="amount_untaxed" sum="Total Untaxed"/>
-        </field>
-        <field name="amount_total" position="after">
-            <field name="currency_id" groups="base.group_multi_currency"/>
+            <field name="amount_untaxed" sum="Total Untaxed" widget="monetary"/>
         </field>
     </field>
 </record>
@@ -48,10 +40,7 @@
     <field name="inherit_id" ref="sale.view_order_tree"/>
     <field name="arch" type="xml">
         <field name="amount_total" position="before">
-            <field name="amount_untaxed" sum="Total Untaxed"/>
-        </field>
-        <field name="amount_total" position="after">
-            <field name="currency_id" groups="base.group_multi_currency"/>
+            <field name="amount_untaxed" sum="Total Untaxed" widget="monetary"/>
         </field>
         <field name="date_order" position="attributes">
             <attribute name="invisible">1</attribute>
@@ -62,7 +51,16 @@
     </field>
 </record>
 
-<!-- TODO : handle date_confirm in search/graph views -->
+<record id="view_sales_order_filter" model="ir.ui.view">
+    <field name="name">usability.sale.order.search</field>
+    <field name="model">sale.order</field>
+    <field name="inherit_id" ref="sale.view_sales_order_filter"/>
+    <field name="arch" type="xml">
+        <filter context="{'group_by':'date_order'}" position="after">
+            <filter string="Order Confirmation Month" name="confirmation_date_groupby" context="{'group_by': 'confirmation_date'}"/>
+        </filter>
+    </field>
+</record>
 
 <record id="procurement_form_view" model="ir.ui.view">
     <field name="name">procurement_usability.procurement.order.form</field>
@@ -75,5 +73,17 @@
     </field>
 </record>
 
-</data>
-</openerp>
+<record id="procurement_group_form_view" model="ir.ui.view">
+    <field name="name">sale_usability.procurement.group.form</field>
+    <field name="model">procurement.group</field>
+    <field name="inherit_id" ref="procurement.procurement_group_form_view"/>
+    <field name="arch" type="xml">
+        <xpath expr="//field[@name='move_type']/.." position="after">
+            <group name="sale" string="Sale Orders">
+                <field name="sale_ids" nolabel="1"/>
+            </group>
+        </xpath>
+    </field>
+</record>
+
+</odoo>
