+++ stock_usability/stock_view.xml
@@ -284,5 +284,18 @@ in the quantities for his reception -->
     <field name="view_id" ref="stock.view_move_tree"/>
 </record>
 
+<record id="procurement_group_form_view" model="ir.ui.view">
+    <field name="name">stock_usability.procurement.group.form</field>
+    <field name="model">procurement.group</field>
+    <field name="inherit_id" ref="procurement.procurement_group_form_view"/>
+    <field name="arch" type="xml">
+        <xpath expr="//field[@name='move_type']/.." position="after">
+            <group name="picking" string="Pickings">
+                <field name="picking_ids" nolabel="1"/>
+            </group>
+        </xpath>
+    </field>
+</record>
+
 </data>
 </openerp>
